Refactor top level class
The EventBot class was doing too much. It was a subclass of the client
that interacts with the Discord API, but it was also providing database
functionality.

I've broken apart the Discord API client and the database functionality
into two distinct classes, both of which are now contained by the new
top level class, EventBot.