Refactor event listing logic to its own module
The EventBot object extends the discord.py client. It is intended to be
used as a way to add additional 'helper' methods that string together
various existing API calls. It violates the SRP to have EventChannel
logic on the EventBot, as our EventChannel has nothing to do with the
Discord API.