Remove query options
While this change means that we will no longer have the option to use
these premade queries with joins, I believe it is the best option right
now as the implementation was not amazing. Doing lazy subquerying is not
the end of the world, and the code ends up looking much nicer.

We can always revisit this in the future if it becomes a concern :)