Recycle stale database connections
In the event that there is no interaction with the database for 8 hours,
the connection will be closed (MySQL default). To prevent this, we will
recycle connections that have not been used in the last hour.

This won't be a problem once more people start using the bot, but
currently it can cause the bot to be unresponsive so we should fix this.