Add dynamic time zone abbreviation
This changes both the input of time zones, as well as how we display
them on events.

We now input the iso time zone string when creating an event. This isn't
a great user experience, as they need to manually type out the iso time
zone string. This will be changing in the future to a more user friendly
input method.

The time zone abbreviation is also determined dynamically, pulling the
time zone name from the time zone aware date time object.