Use DI for on_raw_reaction_add and its children
We have a lot of hard dependencies between classes right now, and I
would really like to get away from this.

Instead of reaching out to our dependencies, we can instead have them
passed down to us. This results service objects that have a much smaller
surface area, which is a big win.

As an added note, this will also allow us to reasonably test our
individual services in the future!