Rewrite the universe
This is a pretty disgusting commit. The underlying change is to move to
a place where we are being explicit about our calls to the database.

We previously were leaving database transactions open for a very long
time, and even left sessions open across multiple service objects.
Essentially, we had no idea what the database was doing.

This commit moves to a place where we load a record, and then we're
done, session commited and closed. Each time we need to talk to the
database, we open a new session. Each session/transaction is open for a
very short period of time, so we don't end up in a place where we are
leaving a transaction open with multiple API calls in between.

To make things worse/better, I also decided to run prettier on the
project while I was in the middle of this refactor. Needless to say, I
wasn't overly concerned about git history when I did all this works.
Woops. Though I'm happy with the end result.