Normalize database credentials
We previosly didn't have variables for the host, post, or database name
in all situations. By allowing all of these to be passed, our app can be
used in more than one context.

I've also omitted the line to create the database, as that's not really
a concern of our application, and can assume that whatever database
we're pointed at already exists.