Add EventSummary model
There's quite a bit of additional information we need to build from our
Event model before we can display it. Much of this work involves
reaching out to Discord's API to get information about Discord users who
have RSVP'd to the event.

Currently this logic lives almost entirely inside our event embed
service. This isn't great, as if we ever need to access this information
in a different context, we've no choice but to copy paste.

To solve this, we can extract all of our "extra" information about an
event model to a standalone entity, allowing us to reuse this code
wherever we please.